move record per min logging to base class
